Rename package to metrictest
Follow Go std lib conventions in test package naming.
