+++ sdk/export/metric/metrictest/test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package test
+package metrictest // import "go.opentelemetry.io/otel/sdk/export/metric/metrictest"
 
 import (
 	"context"
